Isa extensions (#1917)
Adding a new function for ECC
